feat: Add MMR helper allowing simd in similarity (#525)
* feat: Add MMR helper allowing simd in similarity

The main change is that it builds up an `ndarray` of embeddings to
perform similarity on, so that it take advantage of the fact that
`cosine_similarity` can compute M * N similarity values all at once.
In turn, this lets it use `simsimd` more effectively to speed up
similarity computations.

It also moves most of the logic for computing and updating similarity
out of the core MMR loop. This will allow further enhancements to the
MMR traversal (such as considering "fuzzy" edges) to be implemented
more cleanly. This also makes it possible to test more of the logic in
isolation.

This also avoids querying for nodes with incoming tags multiple times.
Once we've discovered the nodes with a given incoming tag that won't
change, so we don't need to re-issue the query.